requirements.txt
Django==5.0.4
django-filter==24.2